[refactor|unstable]: prisma migration, users path without types and proper data flow, only working endpoints

+++ server/src/auth/auth.service.ts
@@ -1,21 +1,24 @@
-import { User } from './../users/users.model';
 import { LoginUserDto } from './../users/dto/login-user.dto';
 import { CreateUserDto } from './../users/dto/create-user.dto';
 import { UsersService } from './../users/users.service';
-import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
+import {
+  Injectable,
+  HttpException,
+  HttpStatus,
+  UnauthorizedException,
+} from '@nestjs/common';
 import * as bcrypt from 'bcryptjs';
 import { v4 } from 'uuid';
-import { FilesService } from '../files/files.service';
 import { UserDto } from '../users/dto/user.dto';
 import { TokensService } from '../tokens/tokens.service';
 import { MailService } from '../mail/mail.service';
 import { UserPassDto } from '../users/dto/user-pass.dto';
+import { User } from '@prisma/client';
 
 @Injectable()
 export class AuthService {
   constructor(
     private readonly usersService: UsersService,
-    private readonly filesService: FilesService,
     private readonly tokensService: TokensService,
     private readonly mailService: MailService,
   ) {}
@@ -29,12 +32,10 @@ export class AuthService {
     const hashPassword = await bcrypt.hash(createUserDto.password, 7);
     const activationLink = v4();
 
-    const user = (await this.usersService.create({
+    const user = await this.usersService.create({
       ...createUserDto,
       password: hashPassword,
-    })) as User & { _id: string };
-
-    this.filesService.makeUserDir(user._id.toString());
+    });
 
     await this.mailService
       .sendMail(
@@ -50,9 +51,9 @@ export class AuthService {
       });
 
     const userDto = new UserDto(user);
-    const userPassDto = new UserPassDto({ _id: user._id, email: user.email });
+    const userPassDto = new UserPassDto({ id: user.id, email: user.email });
     const tokens = this.tokensService.generateTokens({ ...userPassDto });
-    await this.tokensService.saveToken(userDto._id, tokens.refreshToken);
+    await this.tokensService.saveToken(userDto.id, tokens.refreshToken);
 
     return { ...tokens, user: userDto };
   }
@@ -60,7 +61,7 @@ export class AuthService {
   async login(loginUserDto: LoginUserDto) {
     const user = (await this.usersService.getByEmail(
       loginUserDto.email,
-    )) as User & { _id: string };
+    )) as User;
     if (!user) {
       throw new HttpException(
         'The user with such an email is not found',
@@ -77,14 +78,18 @@ export class AuthService {
     }
 
     const userDto = new UserDto(user);
-    const userPassDto = new UserPassDto({ _id: user._id, email: user.email });
+    const userPassDto = new UserPassDto({ id: user.id, email: user.email });
     const tokens = this.tokensService.generateTokens({ ...userPassDto });
-    await this.tokensService.saveToken(userDto._id, tokens.refreshToken);
+    await this.tokensService.saveToken(userDto.id, tokens.refreshToken);
 
     return { ...tokens, user: userDto };
   }
 
   async logout(refreshToken: string) {
+    if (!refreshToken) {
+      throw new UnauthorizedException();
+    }
+
     const token = await this.tokensService.removeToken(refreshToken);
     return token;
   }
@@ -106,15 +111,13 @@ export class AuthService {
       );
     }
 
-    const user = (await this.usersService.getOne(userData._id)) as User & {
-      _id: string;
-    };
+    const user = await this.usersService.getOne(userData.id);
 
     const userDto = new UserDto(user);
-    const userPassDto = new UserPassDto({ _id: user._id, email: user.email });
+    const userPassDto = new UserPassDto({ id: user.id, email: user.email });
 
     const tokens = this.tokensService.generateTokens({ ...userPassDto });
-    await this.tokensService.saveToken(userDto._id, tokens.refreshToken);
+    await this.tokensService.saveToken(userDto.id, tokens.refreshToken);
 
     return { ...tokens, user: userDto };
   }
